Dummy Data File
This file is set up to create dummy data based on random variables.  This makes it easy to create a population of any size.  The system is set up so that the final result will (generally) match the populations ratios of the United States.  The racial categories are based on the EEO codes.  The non-binary population has been set at ~2% so it can be included in the analytics.